.github/workflows/requirements.txt
PyYAML~=6.0.1
requests~=2.32